refactor(nfs): move mount dispatch into the mount package
